Refactor the runtime to be a little more clear. Fixed running in a browser.
