+++ runtime/src/core/grain-module.js
@@ -1,4 +1,3 @@
-import fs from 'fs';
 import { GrainError } from '../errors/errors';
 import { grainToJSVal } from '../utils/utils';
